major update
Well I found Re.Class and made myself a nice class of the local player entity so instead of *(ammo) which is defined ammo = *(int*)(localplayer+offset) we have ourselves a nice localPlayer->ammo. The thing is that I identify the memeory adress, i give it a name and a datatype and the program chreates a vector of chars equal to the size of the next element of the class - current size.