Working entity list, FINALLY
Made the entity list to work and had some fun implementing a function that returns the closest entity to the local player and then you can teleport it to a saved location. I also implemented a function to teleport them all at once to a saved location. So much fun tbh